Only implement necessary parts of a SYCL conventions object
Instead of pulling in all the Vulkan conventions.
